+++ behavex/runner.py
@@ -51,7 +51,7 @@ from behavex.utils import (IncludeNameMatch, IncludePathsMatch, MatchInclude,
                            cleanup_folders, configure_logging,
                            copy_bootstrap_html_generator,
                            create_execution_complete_callback_function,
-                           explore_features, generate_reports,
+                           explore_features, generate_hash, generate_reports,
                            get_json_results, get_logging_level,
                            get_scenario_tags, get_scenarios_instances,
                            get_text, join_feature_reports,
@@ -71,7 +71,6 @@ os.environ.setdefault('EXECUTION_CODE', '1')
 match_include = None
 include_path_match = None
 include_name_match = None
-scenario_lines = {}
 
 
 def main():
@@ -251,7 +250,7 @@ def launch_behavex():
                                                                 feature_filename=None,
                                                                 feature_json_skeleton=None,
                                                                 scenarios_to_run_in_feature=None,
-                                                                scenario_name=None,
+                                                                scenario_line=None,
                                                                 multiprocess=False,
                                                                 config=config,
                                                                 lock=None,
@@ -286,11 +285,10 @@ def launch_behavex():
         results = get_json_results()
         processed_feature_filenames = []
         if results:
-            failures = {}
+            failures = []
             for feature in results['features']:
                 processed_feature_filenames.append(feature['filename'])
                 filename = feature['filename']
-                failures[filename] = []
                 if feature['status'] == 'failed':
                     totals['features']['failed'] += 1
                 elif feature['status'] == 'passed':
@@ -304,7 +302,7 @@ def launch_behavex():
                 for scenario in feature['scenarios']:
                     if scenario['status'] == 'failed':
                         totals['scenarios']['failed'] += 1
-                        failures[filename].append(scenario['name'])
+                        failures.append('{}:{}'.format(filename, scenario['line']))
                         if 'MUTE' not in scenario['tags']:
                             failing_non_muted_tests = True
                     elif scenario['status'] == 'passed':
@@ -316,8 +314,7 @@ def launch_behavex():
             if failures:
                 failures_file_path = os.path.join(get_env('OUTPUT'), global_vars.report_filenames['report_failures'])
                 with open(failures_file_path, 'w') as failures_file:
-                    parameters = create_test_list(failures)
-                    failures_file.write(parameters)
+                    failures_file.write(','.join(failures))
         # Calculates final exit code. execution_codes is 1 only if an execution exception arises
         if isinstance(execution_codes, list):
             execution_failed = True if sum(execution_codes) > 0 else False
@@ -375,23 +372,6 @@ def notify_missing_features(features_path):
                 print_parallel('path.not_found', os.path.realpath(include_path))
 
 
-def create_test_list(test_list):
-    """Create a list of tests to run.
-
-    Args:
-        test_list (dict): Dictionary of features and their scenarios.
-
-    Returns:
-        str: Comma-separated list of test paths.
-    """
-    paths = []
-    sce_lines = get_env('scenario_lines')
-    for feature, scenarios in test_list.items():
-        for scenario_name in scenarios:
-            paths.append('{}:{}'.format(feature, sce_lines[feature][scenario_name]))
-    return ','.join(paths)
-
-
 def create_scenario_line_references(features):
     """Create references for scenario lines in the features.
 
@@ -401,13 +381,10 @@ def create_scenario_line_references(features):
     Returns:
         dict: Updated features with scenario line references.
     """
-    sce_lines = {}
     updated_features = {}
     for feature_path, scenarios in features.items():
         for scenario in scenarios:
             scenario_filename = text(scenario.filename)
-            if scenario_filename not in sce_lines:
-                sce_lines[scenario_filename] = {}
             if global_vars.rerun_failures or ".feature:" in feature_path:
                 feature_without_scenario_line = feature_path.split(":")[0]
                 if feature_without_scenario_line not in updated_features:
@@ -417,27 +394,22 @@ def create_scenario_line_references(features):
                         if scenario_outline_instance.line == int(feature_path.split(":")[1]):
                             if scenario_outline_instance not in updated_features[feature_without_scenario_line]:
                                 updated_features[feature_without_scenario_line].append(scenario_outline_instance)
-                            sce_lines[scenario_filename][scenario_outline_instance.name] = scenario_outline_instance.line
                             break
                 else:
                     if scenario.line == int(feature_path.split(":")[1]):
                         if scenario not in updated_features[feature_without_scenario_line]:
                             updated_features[feature_without_scenario_line].append(scenario)
-                        sce_lines[scenario_filename][scenario.name] = scenario.line
             else:
                 updated_features_path = scenario.feature.filename
                 if updated_features_path not in updated_features:
                     updated_features[updated_features_path] = []
                 if isinstance(scenario, ScenarioOutline):
                     for scenario_outline_instance in scenario.scenarios:
-                        sce_lines[scenario_filename][scenario_outline_instance.name] = scenario_outline_instance.line
                         if scenario_outline_instance not in updated_features[updated_features_path]:
                             updated_features[updated_features_path].append(scenario_outline_instance)
                 else:
-                    sce_lines[scenario_filename][scenario.name] = scenario.line
                     if scenario not in updated_features[updated_features_path]:
                         updated_features[updated_features_path].append(scenario)
-    set_env('scenario_lines', sce_lines)
     return updated_features
 
 
@@ -481,7 +453,7 @@ def launch_by_feature(features,
                                                      feature_filename=serial_feature["feature_filename"],
                                                      feature_json_skeleton=serial_feature["feature_json_skeleton"],
                                                      scenarios_to_run_in_feature=None,
-                                                     scenario_name=None,
+                                                     scenario_line=None,
                                                      multiprocess=True,
                                                      config=ConfigRun(),
                                                      lock=None,
@@ -500,7 +472,7 @@ def launch_by_feature(features,
                                      feature_filename=feature_filename,
                                      feature_json_skeleton=feature_json_skeleton,
                                      scenarios_to_run_in_feature=None,
-                                     scenario_name=None,
+                                     scenario_line=None,
                                      multiprocess=True,
                                      config=ConfigRun(),
                                      lock=lock,
@@ -553,7 +525,7 @@ def launch_by_scenario(features,
                     feature_filename = scenario.feature.filename
                     scenario_information = {"feature_filename": feature_filename,
                                             "feature_json_skeleton": feature_json_skeleton,
-                                            "scenario_name": scenario.name}
+                                            "scenario_line": scenario.line}
                     total_scenarios_to_run[feature_filename] = total_scenarios_to_run.setdefault(feature_filename, 0) + 1
                     if 'SERIAL' in scenario_tags:
                         for key, list_scenarios in serial_scenarios.items():
@@ -587,7 +559,7 @@ def launch_by_scenario(features,
                                                             feature_filename=scen_info["feature_filename"],
                                                             feature_json_skeleton=scen_info["feature_json_skeleton"],
                                                             scenarios_to_run_in_feature=scenarios_to_run_in_feature,
-                                                            scenario_name=scen_info["scenario_name"],
+                                                            scenario_line=scen_info["scenario_line"],
                                                             multiprocess=True,
                                                             config=ConfigRun(),
                                                             lock=None,
@@ -604,13 +576,13 @@ def launch_by_scenario(features,
                 scenarios_to_run_in_feature = total_scenarios_to_run[scenario_information["feature_filename"]]
                 feature_filename = scenario_information["feature_filename"]
                 feature_json_skeleton = scenario_information["feature_json_skeleton"]
-                scenario_name = scenario_information["scenario_name"]
+                scenario_line = scenario_information["scenario_line"]
                 future = process_pool.submit(execute_tests,
                                              features_path=features_path,
                                              feature_filename=feature_filename,
                                              feature_json_skeleton=feature_json_skeleton,
                                              scenarios_to_run_in_feature=scenarios_to_run_in_feature,
-                                             scenario_name=scenario_name,
+                                             scenario_line=scenario_line,
                                              multiprocess=True,
                                              config=ConfigRun(),
                                              lock=lock,
@@ -632,7 +604,7 @@ def execute_tests(
         feature_filename,
         feature_json_skeleton,
         scenarios_to_run_in_feature,
-        scenario_name,
+        scenario_line,
         multiprocess,
         config,
         lock,
@@ -645,7 +617,7 @@ def execute_tests(
         feature_filename (str): Name of the feature file.
         feature_json_skeleton (str): JSON skeleton of the feature.
         scenarios_to_run_in_feature (int): Number of scenarios to run in the feature.
-        scenario_name (str): Name of the scenario.
+        scenario_line (int): Line of the scenario.
         multiprocess (bool): Whether to use multiprocessing.
         config (ConfigRun): Configuration object.
         lock (Lock): Multiprocessing lock.
@@ -667,7 +639,7 @@ def execute_tests(
                                                 multiprocess=multiprocess,
                                                 execution_id=execution_id,
                                                 feature=feature_filename,
-                                                scenario=scenario_name,
+                                                scenario_line=scenario_line,
                                                 config=config)
         except Exception as exception:
             traceback.print_exc()
@@ -682,9 +654,9 @@ def execute_tests(
                                    'features': [json.loads(feature_json_skeleton)],
                                    'steps_definition': []}
                     for skeleton_feature in json_output["features"]:
-                        if scenario_name:
+                        if scenario_line:
                             for skeleton_scenario in skeleton_feature["scenarios"]:
-                                if scenario_name_matching(scenario_name, skeleton_scenario['name']):
+                                if str(skeleton_scenario['line']) == str(scenario_line):
                                     skeleton_scenario['status'] = 'failed'
                                     skeleton_scenario['error_msg'] = get_text('scenario.execution_crashed')
                         else:
@@ -697,16 +669,16 @@ def execute_tests(
                     json_output = {'environment': [], 'features': [], 'steps_definition': []}
             else:
                 json_output = dump_json_results()
-            if scenario_name:
+            if scenario_line:
                 json_output['features'] = filter_feature_executed(json_output,
                                                                   text(feature_filename),
-                                                                  scenario_name)
+                                                                  scenario_line=scenario_line)
                 if len(json_output['features']) == 0 or len(json_output['features'][0]['scenarios']) == 0:
                     # Adding scenario data if the test was removed from the execution (setting it as "Untested")
                     json_output['features'] = [json.loads(feature_json_skeleton)]
                 try:
                     processing_xml_feature(json_output=json_output,
-                                           scenario=scenario_name,
+                                           scenario_line=scenario_line,
                                            feature_filename=feature_filename,
                                            scenarios_to_run_in_feature=scenarios_to_run_in_feature,
                                            lock=lock,
@@ -721,20 +693,20 @@ def execute_tests(
         raise
 
 
-def filter_feature_executed(json_output, filename, scenario_name):
+def filter_feature_executed(json_output, filename, scenario_line):
     """
     Filter the executed feature from the JSON output.
 
     Args:
         json_output (dict): JSON output of the test execution.
         filename (str): Name of the feature file.
-        scenario_name (str): Name of the scenario.
+        scenario_line (str): Line of the scenario.
     """
     for feature in json_output.get('features', '')[:]:
         if feature.get('filename', '') == filename:
             mapping_scenarios = []
             for scenario in feature['scenarios']:
-                if scenario_name_matching(scenario_name, scenario['name']):
+                if str(scenario['line']) == str(scenario_line):
                     mapping_scenarios.append(scenario)
             feature['scenarios'] = mapping_scenarios
             return [feature]
@@ -881,7 +853,7 @@ def remove_temporary_files(parallel_processes, json_reports):
                     print(remove_ex)
 
 
-def processing_xml_feature(json_output, scenario, feature_filename,
+def processing_xml_feature(json_output, scenario_line, feature_filename,
                            scenarios_to_run_in_feature=None, lock=None,
                            shared_removed_scenarios=None):
     """
@@ -911,9 +883,7 @@ def processing_xml_feature(json_output, scenario, feature_filename,
             reported_scenarios = json_output['features'][0]['scenarios']
             executed_scenarios = []
             for reported_scenario in reported_scenarios:
-                reported_name = reported_scenario['name']
-                if reported_name == scenario or ('@' in reported_name and
-                                                 scenario_name_matching(scenario, reported_name)):
+                if reported_scenario['line'] == scenario_line:
                     executed_scenarios.append(reported_scenario)
             json_output['features'][0]['scenarios'] = executed_scenarios
             feature_name = os.path.join(
@@ -1031,7 +1001,7 @@ def _store_tags_to_env_variable(tags):
         set_env_variable('TAGS', '')
 
 
-def _set_behave_arguments(features_path, multiprocess, execution_id=None, feature=None, scenario=None, config=None):
+def _set_behave_arguments(features_path, multiprocess, execution_id=None, feature=None, scenario_line=None, config=None):
     """
     Set the arguments for Behave framework based on the given parameters.
 
@@ -1040,7 +1010,7 @@ def _set_behave_arguments(features_path, multiprocess, execution_id=None, featur
         multiprocess (bool): Whether to use multiprocessing.
         execution_id (str): Execution ID.
         feature (Feature): Feature object.
-        scenario (Scenario): Scenario object.
+        scenario_line (int): Scenario line.
         config (ConfigRun): Configuration object.
 
     Returns:
@@ -1049,24 +1019,10 @@ def _set_behave_arguments(features_path, multiprocess, execution_id=None, featur
     arguments = []
     output_folder = config.get_env('OUTPUT')
     if multiprocess:
-        if not feature:
-            arguments.append(features_path)
-        else:
-            arguments.append(feature)
+        updated_features_path = features_path if not feature else feature
+        updated_features_path = updated_features_path if not scenario_line else "{}:{}".format(updated_features_path, scenario_line)
+        arguments.append(updated_features_path)
         arguments.append('--no-summary')
-        if scenario:
-            outline_examples_in_name = re.findall('<[\\S]*>', scenario)
-            pattern = "(.?--.?@\\d+.\\d+\\s*\\S*)"
-            if bool(re.search(pattern, scenario)):
-                scenario_outline_compatible = '^{}$'.format(re.escape(scenario))
-            else:
-                scenario_outline_compatible = '^{}{}?$'.format(re.escape(scenario), pattern)
-            if outline_examples_in_name:
-                for example_name in outline_examples_in_name:
-                    escaped_example_name = re.escape(example_name)
-                    scenario_outline_compatible = scenario_outline_compatible.replace(escaped_example_name, "[\\S ]*")
-            arguments.append('--name')
-            arguments.append("{}".format(scenario_outline_compatible))
         worker_id = multiprocessing.current_process().name.split('-')[-1]
 
         arguments.append('--outfile')
@@ -1190,26 +1146,6 @@ def set_args_captures(args, args_sys):
     for default_arg in ['capture', 'capture_stderr', 'logcapture']:
         if not getattr(args_sys, 'no_{}'.format(default_arg)):
             args.append('--no-{}'.format(default_arg.replace('_', '-')))
-
-
-def scenario_name_matching(abstract_scenario_name, scenario_name):
-    """
-    Check if the scenario name matches the abstract scenario name (as the scenario might represent a Scenario Outline, with example parameters in name).
-
-    Args:
-        abstract_scenario_name (str): Abstract scenario name
-        scenario_name (str): Scenario name to map against the abstract scenario name.
-
-    Returns:
-        bool: Whether the scenario name matches the abstract scenario name.
-    """
-    outline_examples_in_name = re.findall('<\\S*>', abstract_scenario_name)
-    scenario_outline_compatible = '^{}(.--.@\\d+.\\d+\\s*\\S*)?$'.format(re.escape(abstract_scenario_name))
-    for example_name in outline_examples_in_name:
-        escaped_example_name = re.escape(example_name)
-        scenario_outline_compatible = scenario_outline_compatible.replace(escaped_example_name, "[\\S ]*")
-    pattern = re.compile(scenario_outline_compatible)
-    return pattern.match(scenario_name)
 
 
 def dump_json_results():